webui: move pages components
