+++ webui/src/components/Content/PreTag.tsx
@@ -1,5 +1,6 @@
-import { makeStyles } from '@material-ui/styles';
 import React from 'react';
+
+import { makeStyles } from '@material-ui/styles';
 
 const useStyles = makeStyles({
   tag: {
