+++ webui/src/pages/bug/Bug.graphql
