+++ webui/src/pages/bug/LabelChange.tsx
@@ -1,9 +1,10 @@
-import { makeStyles } from '@material-ui/core/styles';
 import React from 'react';
 
-import Author from '../components/Author';
-import Date from '../components/Date';
-import Label from '../components/Label';
+import { makeStyles } from '@material-ui/core/styles';
+
+import Author from 'src/components/Author';
+import Date from 'src/components/Date';
+import Label from 'src/components/Label';
 
 import { LabelChangeFragment } from './LabelChangeFragment.generated';
