+++ webui/src/pages/bug/LabelChangeFragment.graphql
@@ -1,4 +1,4 @@
-#import "../components/fragments.graphql"
+#import "../../components/fragments.graphql"
 
 fragment LabelChange on LabelChangeTimelineItem {
   date
