+++ webui/src/pages/bug/MessageCommentFragment.graphql
@@ -1,4 +1,4 @@
-#import "../components/fragments.graphql"
+#import "../../components/fragments.graphql"
 
 fragment AddComment on AddCommentTimelineItem {
   createdAt
