+++ webui/src/pages/bug/MessageCreateFragment.graphql
@@ -1,4 +1,4 @@
-#import "../components/fragments.graphql"
+#import "../../components/fragments.graphql"
 
 fragment Create on CreateTimelineItem {
   createdAt
