+++ webui/src/pages/bug/SetStatusFragment.graphql
@@ -1,4 +1,4 @@
-#import "../components/fragments.graphql"
+#import "../../components/fragments.graphql"
 
 fragment SetStatus on SetStatusTimelineItem {
   date
