+++ webui/src/pages/bug/SetTitleFragment.graphql
@@ -1,4 +1,4 @@
-#import "../components/fragments.graphql"
+#import "../../components/fragments.graphql"
 
 fragment SetTitle on SetTitleTimelineItem {
   date
