+++ webui/src/pages/bug/TimelineQuery.tsx
@@ -1,5 +1,6 @@
-import CircularProgress from '@material-ui/core/CircularProgress';
 import React from 'react';
+
+import CircularProgress from '@material-ui/core/CircularProgress';
 
 import Timeline from './Timeline';
 import { useTimelineQuery } from './TimelineQuery.generated';
